Version 2.0
Added a two-body and three-body interaction with key values like velocity and mass modifiable by a slider. Because of varying options now, I made a simple main menu that allow scenes to be traversed.

Two body was possible with the generic gravitational force script (SatelliteBehaviour) that I used to make simple one-body motion, but it was incredibly challenging to make a generic force calculator for three or more body. Specifically, it was difficult because I don't know how to make a variable size array to store body gameobjects, and even if I know how to I cannot guarantee that when there are too many bodies the calculation could be completed on time (because there would be n-1 amount of calculations for each of the n bodies,) So I made a version specifically for 3 body for now.

I learnt that this specific approach that I am taking is called the Euler method. which is different from the Verlet integration I used before in BPhO. I also learnt that Euler method have numerical error that gets bigger when the timestep gets larger (because each step is a straight line, not the curve), and will eventually push the satellite out of its expected orbit. So I think for the sake of both physical-and-mathematical rigor and not losing accuracy with variable timestep so that I can work on generic N-body force calculation I will work on Verlet-ising my current codes as my next goal.

Testing the three body simulation, I noticed that it seems to always simplify into pseudo-two-body state where two close bodies orbit each other with net velocity heading away from the third body. I have yet to find a reasonable explanation to this.