Version 3.0
When I ran the previous Euler orbit mechanism to run for a long time, I noticed that it deviated from the orbit, specifically going towards the planet. After a brief research I found out that this method is inherently prone to error as it only accounted for the acceleration at the beginning of motion happening in dt, not accounting for change in velocity between dt. This meant that, in my application specifically, the satellite will move inward over time as it does not account for forces pulling it in between dts, hence leading to overall decrease in mechanical energy of the system. I thought taking changing acceleration between dts into account would make the computation more accurate, and predicted that Verlet integration method I learnt from BPhO Computational challenge would be able to solve the problem. Averaging out the velocity between two dts did indeed yield a stable orbit.

So I implemented Verlet integration and used it to make two, three, and four body simulations. In previous Euler version I observed that in multi body motion one planet would slingshot from the other when they got too close to each other. Verlet didn't show noticeable difference in low accuracy setting (dt about 0.01s)  compared to Euler but in high accuracy setting (dt about 0.001s or lower) bodies did not slingshot themselves and actually got to orbit each other, making multi body simulations much more stable. Because acceleration is calculated only when dt is ticked, slower approaching satellite will have more dt ticked than faster departing satellite per distance, hence departing satellites will not have enough force applied to them compared to real life. Lowering dt solved this problem by allowing more forces to be applied to the departing satellite per distance.

My initial reason of using Verlet was to reduce error in orbital movement, and I wanted to confirm that intuition. To find it out, I made a python program to calculate the change of mechanical energy of the Euler integration and compared it with that of Verlet. It turned out that Euler method's mechanical energy was not conserved, but rather decreased over time. Verlet's ME, on the other hand, did not changed up to 7 significant figures. This confirmed that Verlet do form more stable orbit than Euler in a long run because it conserves mechanical energy.

Lower dt meant more accurate orbital simulation (as shown by multi body simulation) but more computation needed to carry out. Because of that, for optimisation purpose, I thought of adapting a variable dt depending on the distance between the planet and the satellite: closer they are dt will be lower for accurate computation and farther they are dt will be higher because long distance means less force applied and less drastic changes to the velocity and position. It did not work, however; if dt was changed, the acceleration and velocity seems to scale proportional with it and break the simulation. Further research told me that it is because Verlet is 'time symmetrical', meaning that change in position made in forward direction of time should be the exact opposite if the time is to flow backwards. If dt was to decrease, the distance moved in a single dt from that moment wouldn't be the same as higher dt it used to move in reference with, and will start to move in orbit as if it started with that small distance moved per dt, hence having smaller orbit. This was exactly what was observed in my variable dt testing; satellite approaching planet had smaller orbit due to their decreased dt while outgoing satellites' velocity exploded due to their increased dt.